# stats: Measurement doesn't need to be a pointer.