--- a/stats/measure_int64.go
+++ b/stats/measure_int64.go
@@ -39,6 +39,6 @@ func (m *MeasureInt64) Unit() string {
 
 // M creates a new int64 measurement.
 // Use Record to record measurements.
-func (m *MeasureInt64) M(v int64) *Measurement {
-	return &Measurement{m: m, v: v}
+func (m *MeasureInt64) M(v int64) Measurement {
+	return Measurement{m: m, v: v}
 }