Also added a simple benchmark to show that we reduce allocations by not using pointers.

Before:

```
go test -bench=BenchmarkRecord -benchmem -benchtime=3s
goos: darwin
goarch: amd64
pkg: go.opencensus.io/stats
BenchmarkRecord-8   	 2000000	      2816 ns/op	     600 B/op	      24 allocs/op
PASS
ok  	go.opencensus.io/stats	8.939s
```

After:

```
go test -bench=BenchmarkRecord -benchmem -benchtime=3s
goos: darwin
goarch: amd64
pkg: go.opencensus.io/stats
BenchmarkRecord-8   	 2000000	      2393 ns/op	     520 B/op	      14 allocs/op
PASS
ok  	go.opencensus.io/stats	7.647s
```